Enable/Disable camera working reliably
The way disabling the camera didn't work reliably since we have a self-preview in our videoroom. The self-preview of the local stream stayed black. Using StreamHelpers also for the camera seems to work more stable and fixed the issue. Our code drifted quite a bit but if you agree that this is a better solution I think it should be merged. Please double check functionality before merging.